Made HashContainer.currentLoad() to return double instead of float

+++ lib/api/src/main/java/net/openhft/collect/HashContainer.java
@@ -36,7 +36,7 @@ public interface HashContainer extends Container {
      *
      * @return fullness of the hash
      */
-    float currentLoad();
+    double currentLoad();
 
 
     /**
