+++ lib/impl/src/main/javaTemplates/net/openhft/collect/impl/hash/MutableDHash.java
@@ -117,9 +117,8 @@ public abstract class MutableDHash extends AbstractContainer implements DHash {
     }
 
     @Override
-    public final float currentLoad() {
-        // Division in double to minimize precision loss
-        return (float) (((double) (size + removedSlots)) / (double) capacity());
+    public final double currentLoad() {
+        return ((double) (size + removedSlots)) / (double) capacity();
     }
